fix: Reference error of alist_to_js_obj
