+++ src/library/webscheme_lib.js
@@ -7,7 +7,7 @@ import { define_libfunc, alias_libfunc, define_syntax, define_scmfunc,
          assert_function, assert_closure, assert_procedure, assert_date, deprecate } from "./infra.js"; 
 import Console from "../system/console.js"
 import Interpreter  from "../system/interpreter.js"
-import { Pair } from "../system/pair.js"
+import { Pair, alist_to_js_obj } from "../system/pair.js"
 import Pause from "../system/pause.js"
 import { BiwaSymbol, Sym } from "../system/symbol.js"
